improved description and assembly of the LVPP example for the obstacle problem
